pyproject.toml
[tool.poetry]
name = "taskplanner"
version = "0.1.0"
description = ""
authors = ["Andrew-715 <kutuzkin98@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django-environ = "^0.10.0"
psycopg2-binary = "^2.9.6"
django = "^4.2"
gunicorn = "^20.1.0"
djangorestframework = "^3.14.0"
social-auth-app-django = "^5.2.0"
django-filter = "^23.2"
django-stubs = {version = "^4.2.1", extras = ["compatible-mypy"]}

[tool.mypy]
exclude = '.venv'
show_error_codes = true
ignore_missing_imports = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
warn_unused_ignores = true
warn_no_return = false
no_implicit_optional = true
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "todolist.settings"

[[tool.mypy.overrides]]
module = [
    'tests.*',
    'manage',
]
ignore_errors = true


[build-system]
requires = ["poetry-core>=1.4.2"]
build-backend = "poetry.core.masonry.api"
